@@ -71,7 +71,7 @@ public:
            unsigned long radius = 0,
            const PreferredSide& side = PreferredSide::EITHER,
            const SearchFilter& search_filter = SearchFilter(),
-           boost::optional<int8_t> preferred_layer = {});
+           std::optional<int8_t> preferred_layer = {});
 
   /**
    * equality.
@@ -89,8 +89,8 @@ public:
   std::string name_;
   std::string street_;
 
-  boost::optional<std::string> date_time_;
-  boost::optional<float> heading_;
+  std::optional<std::string> date_time_;
+  std::optional<float> heading_;
 
   // try to find candidates who are reachable from/to this many or more nodes
   // if a given candidate edge is reachable to/from less than this number of nodes its considered to
@@ -111,9 +111,9 @@ public:
   SearchFilter search_filter_;
 
   // coordinates of the location as used for altering the side of street
-  boost::optional<midgard::PointLL> display_latlng_;
+  std::optional<midgard::PointLL> display_latlng_;
 
-  boost::optional<int8_t> preferred_layer_;
+  std::optional<int8_t> preferred_layer_;
 
 protected:
 };
