@@ -1,10 +1,12 @@
-#include "sif/dynamiccost.h"
+#include <boost::optional>
+#include <utility>
 
 #include "baldr/graphconstants.h"
 #include "midgard/util.h"
 #include "proto_conversions.h"
 #include "sif/autocost.h"
 #include "sif/bicyclecost.h"
+#include "sif/dynamiccost.h"
 #include "sif/motorcyclecost.h"
 #include "sif/motorscootercost.h"
 #include "sif/nocost.h"
@@ -12,7 +14,6 @@
 #include "sif/transitcost.h"
 #include "sif/truckcost.h"
 #include "worker.h"
-#include <utility>
 
 using namespace valhalla::baldr;
 
