@@ -1,7 +1,3 @@
-#include "baldr/rapidjson_utils.h"
-#include <boost/format.hpp>
-#include <boost/optional.hpp>
-#include <boost/property_tree/ptree.hpp>
 #include <cmath>
 #include <cstdint>
 #include <cstdlib>
@@ -10,10 +6,13 @@
 #include <string>
 #include <vector>
 
-#include "config.h"
+#include "baldr/rapidjson_utils.h"
+#include <boost/format.hpp>
+#include <boost/property_tree/ptree.hpp>
 
 #include "baldr/graphreader.h"
 #include "baldr/pathlocation.h"
+#include "config.h"
 #include "loki/worker.h"
 #include "midgard/logging.h"
 #include "odin/directionsbuilder.h"
@@ -152,8 +151,7 @@ int main(int argc, char* argv[]) {
   rapidjson::read_json(config.c_str(), pt);
 
   // configure logging
-  boost::optional<boost::property_tree::ptree&> logging_subtree =
-      pt.get_child_optional("thor.logging");
+  auto logging_subtree = pt.get_child_optional("thor.logging");
   if (logging_subtree) {
     auto logging_config =
         valhalla::midgard::ToMap<const boost::property_tree::ptree&,
